added a decorator run_if_not_pickled
If result of the function are already pickled, then load the pickle and
return the result, else run the function and dump the results to a given
pickle file.

Function is run only once to do a computation and cached results are
returned. May be I should renamed the decorator to 'cached'.